SDL: Move EGL init code into SDLGLGraphicsContext.
It should really be cleaned up, but I'm not sure if EGL_Open can be
moved later and don't currently have a setup to test, so I'm just
cargo culting.
